festival: Store and display attendance info for registrations
This was already used for a long time on koppelting.org, but was
previously committed on a separate branch (since not all festivals might
use it). However, in the interest of keeping a single, clean version of
the code, this is now committed to master as well (it can always be
made configurable later).