festival: Allow setting a description for a festival page
This will be shown above the lineup.

This was previously implemented as a local change for Koppelting, but
now added to the master branch. Previously, this would support custom
HTML, but since there is no system for proper validation of that yet,
this now changed to text only (wrapped in a classed paragraph so the
site CSS can style the text).

This also changes the lineup generation to use domdocument earlier
instead of first concatenating a lot of HTML.