festival: Update to Mollie API v2
The v1 API was deprecated and is now shut down, so this restores payment
functionality.

This replaces the Mollie-provided API client with a very minimal
(payments only) custom implementation (since the API is really just
a simple REST API anyway).

This prevents having to include a ton of files (the latest Mollie PHP
client is 600+ files, the latest version that supports PHP7 is 200+ or
so).